chore(Admin) - Refactored code in folder services
